+++ src/utils/state/index.js
@@ -1,22 +1,14 @@
 import store from './store';
 import { dispatch } from './dispatcher';
-import { connect, provide } from './connect';
 import { dehydrate, rehydrate, hotRehydrate } from './hydrate';
 import { fetchData, fetchDataOnLocationMatch } from './fetch';
-import ContextProvider from './ContextProvider';
-import Context from './Context';
-
 
 export {
   store,
   dispatch,
-  connect,
-  provide,
   dehydrate,
   rehydrate,
   hotRehydrate,
   fetchData,
   fetchDataOnLocationMatch,
-  ContextProvider,
-  Context,
 };