+++ src/web/App.jsx
@@ -1,12 +1,9 @@
 import React, { Component } from 'react';
 import { Router } from 'react-router';
+import { Provider } from 'mobx-react';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import context from '~/src/shared/context';
 
-const ContextProvider = context.getProvider();
-
-export default
-class App extends Component {
+export default class App extends Component {
 
   static propTypes = {
     store: React.PropTypes.object,
@@ -19,9 +16,9 @@ class App extends Component {
   render() {
     return (
       <MuiThemeProvider muiTheme={this.props.store.ui.getMui()}>
-        <ContextProvider context={{ store: this.props.store }}>
+        <Provider store={this.props.store}>
           <Router routes={this.props.routes} history={this.props.history} />
-        </ContextProvider>
+        </Provider>
       </MuiThemeProvider>
     );
   }